dynamic-pos/services/auth/requirements.txt
Django>=4.2,<5
djangorestframework>=3.14
djangorestframework-simplejwt>=5.2
psycopg2-binary
django-cors-headers
celery[redis]>=5.2
kombu
requests
python-dotenv
pycryptodome
dj-database-url>=2.0.0
celery
redis
gunicorn
cryptography>=39.0.0
pyjwt[crypto]>=2.6.0
jwcrypto
django-extensions
whitenoise
drf-spectacular
